+++ lobster/tools/trlc/lobster_trlc_config.py
@@ -5,10 +5,10 @@ from typing import List
 import yamale
 import yaml
 
-from lobster.items import Requirement
+from lobster.common.items import Requirement
+from lobster.common.multi_file_input_config import Config
 from lobster.tools.trlc.conversion_rule import ConversionRule
 from lobster.tools.trlc.to_string_rules import ToStringRules
-from lobster.multi_file_input_config import Config
 
 
 @dataclass