+++ tests_system/lobster_json/test_name_attribute.py
@@ -1,8 +1,9 @@
 from os import makedirs
 import shutil
-from .lobsterjsonsystemtestcasebase import LobsterJsonSystemTestCaseBase
-from .lobsterjsonasserter import LobsterJsonAsserter
-from ..asserter import Asserter
+from tests_system.lobster_json.\
+    lobsterjsonsystemtestcasebase import LobsterJsonSystemTestCaseBase
+from tests_system.lobster_json.lobsterjsonasserter import LobsterJsonAsserter
+from tests_system.asserter import Asserter
 
 
 class JsonExtensionTest(LobsterJsonSystemTestCaseBase):