+++ tests_system/lobster_json/test_tag_attribute.py
@@ -1,6 +1,7 @@
-from .lobsterjsonsystemtestcasebase import LobsterJsonSystemTestCaseBase
-from .lobsterjsonasserter import LobsterJsonAsserter
-from ..asserter import Asserter
+from tests_system.lobster_json.\
+    lobsterjsonsystemtestcasebase import LobsterJsonSystemTestCaseBase
+from tests_system.lobster_json.lobsterjsonasserter import LobsterJsonAsserter
+from tests_system.asserter import Asserter
 
 
 class JsonTagAttributeTest(LobsterJsonSystemTestCaseBase):