+++ tests_system/lobster_meta_data_tool_base/test_meta_data_tool_base.py
@@ -2,13 +2,18 @@ import os
 from tempfile import NamedTemporaryFile
 from dataclasses import dataclass
 from typing import Tuple, Type
-from .lobster_meta_data_tool_base_system_test_case_base import (
-    LobsterMetaDataToolBaseSystemTestCaseBase,
-)
-from .lobster_meta_data_tool_base_asserters import (
-    SpecificAsserter, HelpAsserter, VersionAsserter, IMPLEMENTATION_MESSAGE
-)
-from ..asserter import Asserter
+from tests_system.lobster_meta_data_tool_base.\
+    lobster_meta_data_tool_base_system_test_case_base import (
+        LobsterMetaDataToolBaseSystemTestCaseBase
+    )
+from tests_system.lobster_meta_data_tool_base.\
+    lobster_meta_data_tool_base_asserters import (
+        SpecificAsserter,
+        HelpAsserter,
+        VersionAsserter,
+        IMPLEMENTATION_MESSAGE
+    )
+from tests_system.asserter import Asserter
 
 
 class ToolBaseTest(LobsterMetaDataToolBaseSystemTestCaseBase):