+++ tests_system/lobster_trlc/test_inputs_from_directory.py
@@ -1,7 +1,8 @@
 import os
 import shutil
-from .lobster_system_test_case_base import LobsterTrlcSystemTestCaseBase
-from ..asserter import Asserter
+from tests_system.lobster_trlc.lobster_system_test_case_base import (
+    LobsterTrlcSystemTestCaseBase)
+from tests_system.asserter import Asserter
 
 
 class InputFromDirectory(LobsterTrlcSystemTestCaseBase):