+++ tests_system/lobster_trlc/test_inputs_zero.py
@@ -1,5 +1,6 @@
-from .lobster_system_test_case_base import LobsterTrlcSystemTestCaseBase
-from ..asserter import Asserter
+from tests_system.lobster_trlc.lobster_system_test_case_base import (
+    LobsterTrlcSystemTestCaseBase)
+from tests_system.asserter import Asserter
 
 
 class ZeroInputTest(LobsterTrlcSystemTestCaseBase):